modify type and repo

+++ ppdiffusers/examples/inference/text_to_image_generation-flux_controlnet_img2img.py
@@ -20,9 +20,9 @@ from ppdiffusers.pipelines import FluxControlNetImg2ImgPipeline
 
 
 controlnet = FluxControlNetModel.from_pretrained("InstantX/FLUX.1-dev-Controlnet-Canny", 
-                                                 paddle_dtype=paddle.float16)
+                                                 paddle_dtype=paddle.bfloat16)
 pipe = FluxControlNetImg2ImgPipeline.from_pretrained(
-    "black-forest-labs/FLUX.1-dev", controlnet=controlnet, paddle_dtype=paddle.float16, low_cpu_mem_usage=True, map_location="cpu",
+    "black-forest-labs/FLUX.1-dev", controlnet=controlnet, paddle_dtype=paddle.bfloat16, low_cpu_mem_usage=True, map_location="cpu",
 )
 
 control_image = load_image("https://huggingface.co/InstantX/SD3-Controlnet-Canny/resolve/main/canny.jpg")
